feat: add `@unocss/utils` package
