+++ packages/preset-mini/src/_utils/index.ts
@@ -1,5 +1,5 @@
-export * from './colors'
 export * from './mappings'
 export * from './handlers'
 export * from './variants'
 export * from './utilities'
+export * from '@unocss/utils'
