+++ packages/utils/README.md
@@ -1,0 +1,7 @@
+# @unocss/utils
+
+ Utilities for [UnoCSS](https://github.com/unocss/unocss).
+ 
+## License
+
+MIT License &copy; 2021-PRESENT [Anthony Fu](https://github.com/antfu)
