+++ tsconfig.json
@@ -51,6 +51,7 @@
       "@unocss/webpack": ["./packages/webpack/src/index.ts"],
       "@unocss/astro": ["./packages/astro/src/index.ts"],
       "@unocss/postcss": ["./packages/postcss/src/index.ts"],
+      "@unocss/utils": ["./packages/utils/src/index.ts"],
       "unocss": ["./packages/unocss/src/index.ts"],
       "unocss/vite": ["./packages/unocss/src/vite.ts"]
     }
